Refactor LinkedQuestion components
