+++ client/src/components/LinkedQuestion/LinkedQuestionOption/LinkedQuestionOption.svelte
