+++ client/src/components/index.js
@@ -4,7 +4,6 @@ export {Footer} from './Footer';
 export {Header} from './Header';
 export {Mathjax} from './Mathjax';
 export {LinkedQuestion} from './LinkedQuestion';
-export {LinkedQuestionOption} from './LinkedQuestionOption';
 export {Loading} from './Loading';
 export {MultipleChoiceAnswer} from './MultipleChoiceAnswer';
 export {MultipleChoiceQuestion} from './MultipleChoiceQuestion';
